pyproject.toml
# File: pyproject.toml

[project]
name = "pibit-prompt-validator"
version = "0.1.0"
description = "A Python module to update and validate a directory of prompts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.scripts]
prompt-validator = "prompt_validator.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# --- ADD THIS SECTION ---
[tool.setuptools]
packages = ["prompt_validator"]